+++ crates/teloxide/src/dispatching.rs
@@ -227,4 +227,4 @@ pub use dispatcher::{Dispatcher, DispatcherBuilder, UpdateHandler};
 pub use distribution::DefaultKey;
 pub use filter_ext::{MessageFilterExt, UpdateFilterExt};
 pub use handler_description::DpHandlerDescription;
-pub use handler_ext::{filter_command, HandlerExt};
+pub use handler_ext::{filter_command, filter_mention_command, HandlerExt};